requirements.txt
web3==7.14.1
requests==2.32.5
python-dotenv==1.2.1
